feat: migrate xray core to sing-box, align with mmwx-pro UI
- Replace xray kernel with sing-box (singboxops, sshclient packages)
- Add securechan compatibility layer (mmwx_compat.go) for legacy Ed25519 API
- Complete SystemConfig/UserSettings SQL fields for mmwx-pro frontend
- Add 17 system_config columns: sub info nodes, login rate limit, brute force protection, sub rate limit, skip_local_ip, block_unknown_subscription_ua
- Add 4 user_settings columns: enable_probe_binding, template_version, default_surge/loon_template_filename
- Add migrateTemplateVersionFromBool migration (use_new_template_system -> template_version)
- Fix RequireToken 3-arg signature (27 call sites)
- Add NewPackagesHandler RESTful adapter
- Deploy verified on 119.28.226.13: service active, HTTP 200, DB migration OK

+++ internal/handler/ddns_admin.go
@@ -67,7 +67,7 @@ func (h *DDNSAdminHandler) handleStatus(w http.ResponseWriter, r *http.Request, 
 	}
 	// 拼一个 DNS provider 名(若已配置 + 有匹配)给 Tooltip 显示用
 	providerName := ""
-	providerID := server.DDNSProviderID
+	providerID := int64(server.DDNSProviderID)
 	if providerID == 0 && server.DDNSEnabled {
 		// 自动模式 — 查匹配证书
 		if cert, cerr := h.repo.FindCertificateForDomain(r.Context(), server.PullAddress); cerr == nil {
